Correct issue #19
When a valid Unicode character is not defined in our tokenizer char
table, backoff to a mapping between Unicode character classes as defined
by Qt and our generic class names.

We should also have a mean to define the mapping in our tokenizer char
table. I will open an new issue for that.