lima_linguisticdata: add cmake policy CMP0046
set to OLD as some dependencies are dynamically created

$ cmake --help-policy CMP0046
CMP0046
-------

Error on non-existent dependency in add_dependencies.

CMake 2.8.12 and lower silently ignored non-existent dependencies
listed in the ``add_dependencies()`` command.

The OLD behavior for this policy is to silently ignore non-existent
dependencies. The NEW behavior for this policy is to report an error
if non-existent dependencies are listed in the ``add_dependencies()``
command.

This policy was introduced in CMake version 3.0.
CMake version 3.1.3 warns when the policy is not set and uses
OLD behavior.  Use the cmake_policy command to set it to OLD or
NEW explicitly.
$ cmake --help-policy CMP0046
CMP0046
-------