Corrects "Empty data for vertex" bug
When a wrong micro-category was used in the definition of a Modex, a wrong null value was set to the newly created token. This commit writes an error message and avoid to add the wrong value. If there is no other micro associated to this entity, an exception is thrown later on when trying to use the empty list.