Allow to use temp files instead of suffix on input
This change allow to create tempoary files and transmit its name in the
analysis metadata to transfer written files from one process unit to
another. Before, we had to use suffixes on the input filename, thus
adding somtimes useless files in the source dir.

The name of the file is stored in a metadata whose name is set in the
process units configuration parameter 'temporaryFileMetadata'.