Correction in CreateSpecificEntity: restart rules after newly created node
correction for issue #44 on github: when a new node is created for a specific entity, the application of the rules restarted on this node. This behavior produces some unexpected results when applying the rules, but was also exploited in some rules: this correction goes with some changes in the rules.