Add CONLL output table display
We can now visualize CONLL format output in a table. A ComboBox has been added
to switch between this and raw text. More view choices to be added in this
drop-down box. This is a part of the new menu for the results. (Maybe we could
add more options in this menu, like exporting, hiding/showing the results tab,
etc.)