Improve named entity type errors handling
Errors related to the absence of an entity in a given group where often
reported only on program exit and with a cryptic group numeric id. This
patch use the group name as soon as possible and tries to intercept the
error at call site to show where the problem occurred.