Revert some modif of 39b64a55 : exception in ViterbiPosTagger::performViterbiOnStepData()
Reason : There should be no constraint on the length of StepDataVector.

For the record: we observed of an infinite loop in LIMA while processing French text on long
sequences of digits in a rulebased semantic pipeline: Domain specific modex,
syntaxic analysis, compound terms.
The erroneous behaviour has been tracked down to the SyntacticAnalyzerNoChains which is
in fact considered obsolete. Commit ec08980 removed all mentions of SyntacticChains
of Lima rulebased.